pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foldermix"
version = "0.1.23"
description = "Pack a folder into a single LLM-friendly context file"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
keywords = ["llm", "context", "packing", "cli"]
dependencies = [
    "click>=8",
    "typer>=0.9",
    "pathspec>=0.12",
    "rich>=13.0",
    "tomli>=2.0 ; python_version < '3.11'",
]

[project.optional-dependencies]
pdf = ["pypdf>=3.0"]
ocr = ["pypdf>=3.0", "pypdfium2>=4.30", "rapidocr-onnxruntime>=1.3"]
office = ["python-docx>=1.0", "openpyxl>=3.1", "python-pptx>=0.6"]
markitdown = ["markitdown>=0.0.1"]
docs = ["mkdocs>=1.6"]
all = [
    "pypdf>=3.0",
    "pypdfium2>=4.30",
    "rapidocr-onnxruntime>=1.3",
    "python-docx>=1.0",
    "openpyxl>=3.1",
    "python-pptx>=0.6",
    "tqdm>=4.0",
]
dev = ["hypothesis>=6.0", "pytest>=7.0", "pytest-cov", "ruff>=0.4", "Pillow>=10.0"]
mutation = ["mutmut>=2.4"]

[project.scripts]
foldermix = "foldermix.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["foldermix"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-vv -ra --cov=foldermix --cov-branch --cov-report=term-missing:skip-covered --cov-fail-under=98"
filterwarnings = ["error"]
markers = [
    "integration: end-to-end and snapshot-style tests",
    "slow: slower tests that may be skipped in smoke runs",
]
strict_config = true
strict_markers = true
testpaths = ["tests"]

[tool.mutmut]
paths_to_mutate = [
    "foldermix/scanner.py",
    "foldermix/packer.py",
    "foldermix/writers/base.py",
    "foldermix/writers/markdown_writer.py",
    "foldermix/writers/xml_writer.py",
    "foldermix/writers/jsonl_writer.py",
]
runner = "python -m pytest -q -o addopts= tests/test_scanner.py tests/test_scanner_edge.py tests/test_scanner_properties.py tests/test_packer.py tests/test_packer_edges.py tests/test_writers.py tests/test_writers_edge.py"
tests_dir = ["tests"]
